(CAT-1733) Unsupport PDK-Rubocop
We have decided to remove support from pdk-rubocop as it is a tool that
has not seen proper implementation yet, nor do we have current plans to
make use of it.

This commit addressed the README file of the tool to inform users of
its new status, as well as announcing its future deprecation.